feat: 実装クラスを実際に使用しているモジュールに移動
